More restructuring the project into different folders
